chore: move {tui,menu}.{c,h} into tui/ subdir
